Moving transit tests to routing_common.
