+++ generator/generator_tests/CMakeLists.txt
@@ -21,9 +21,6 @@ set(
   srtm_parser_test.cpp
   tag_admixer_test.cpp
   tesselator_test.cpp
-  transit_graph_test.cpp
-  transit_test.cpp
-  transit_tools.hpp
   triangles_tree_coding_test.cpp
   types_helper.hpp
   ugc_test.cpp
